Moved hydrator tests out
